Move PNG handling to src/engine/gfx
This way, it's no longer compiled into the server.
