+++ src/engine/client/backend/opengl/backend_opengl3.cpp
@@ -15,7 +15,7 @@
 #include <engine/client/backend/opengl/opengl_sl.h>
 #include <engine/client/backend/opengl/opengl_sl_program.h>
 
-#include <engine/shared/image_manipulation.h>
+#include <engine/gfx/image_manipulation.h>
 
 #include <engine/client/backend/glsl_shader_compiler.h>
