+++ src/engine/client/backend/vulkan/backend_vulkan.cpp
@@ -5,9 +5,9 @@
 #include <engine/client/backend/backend_base.h>
 #include <engine/client/backend_sdl.h>
 #include <engine/client/graphics_threaded.h>
+#include <engine/gfx/image_manipulation.h>
 #include <engine/graphics.h>
 #include <engine/shared/config.h>
-#include <engine/shared/image_manipulation.h>
 #include <engine/storage.h>
 
 #include <base/log.h>
