+++ src/engine/gfx/image_loader.cpp
