+++ src/engine/gfx/image_manipulation.cpp
