+++ src/tools/map_convert_07.cpp
@@ -3,13 +3,12 @@
 
 #include <base/logger.h>
 #include <base/system.h>
+#include <engine/gfx/image_loader.h>
 #include <engine/graphics.h>
 #include <engine/shared/datafile.h>
 #include <engine/storage.h>
 #include <game/gamecore.h>
 #include <game/mapitems.h>
-
-#include <engine/shared/image_loader.h>
 /*
 	Usage: map_convert_07 <source map filepath> <dest map filepath>
 */
