+++ src/tools/map_replace_image.cpp
@@ -3,12 +3,11 @@
 
 #include <base/logger.h>
 #include <base/system.h>
+#include <engine/gfx/image_loader.h>
 #include <engine/graphics.h>
 #include <engine/shared/datafile.h>
 #include <engine/storage.h>
 #include <game/mapitems.h>
-
-#include <engine/shared/image_loader.h>
 /*
 	Usage: map_replace_image <source map filepath> <dest map filepath> <current image name> <new image filepath>
 	Notes: map filepath must be relative to user default teeworlds folder
